tons of shit

+++ src/enemy.rs
@@ -5,10 +5,9 @@ use crate::player::PlayerHealth;
 
 use crate::GameState;
 
+use crate::collision::FlashingTimer;
 use crate::player::PlayerStats;
 use crate::utils::YSort;
-use crate::collision::FlashingTimer;
-
 
 use bevy::prelude::*;
 use rand::Rng;
@@ -18,13 +17,21 @@ pub struct EnemyPlugin<S: States> {
 
 impl<S: States> Plugin for EnemyPlugin<S> {
     fn build(&self, app: &mut App) {
-        app.insert_resource(EnemyTimer(Timer::from_seconds(0.05, TimerMode::Repeating)));
+        app.insert_resource(EnemyTimer(Timer::from_seconds(0.8, TimerMode::Repeating)));
         app.add_systems(
             Update,
-
-            (chase_player, spawn_enemy, wiggle, y_sort, kill_dead_enemies, unfreeze, extinguish, vunerable_tickdown)
-                .run_if(in_state(self.state.clone())).run_if(in_state(GameState::Playing)),
-
+            (
+                chase_player,
+                spawn_enemy,
+                wiggle,
+                y_sort,
+                kill_dead_enemies,
+                unfreeze,
+                extinguish,
+                vunerable_tickdown,
+            )
+                .run_if(in_state(self.state.clone()))
+                .run_if(in_state(GameState::Playing)),
         );
         app.add_systems(OnExit(self.state.clone()), clean_up_enemies);
     }
@@ -43,11 +50,12 @@ pub struct EnemyHealth {
 }
 #[derive(Component)]
 pub struct EnemyXp {
-    pub xp: u32,
+    pub xp: f32,
 }
 #[derive(Component)]
 pub struct ChasePlayer {
     pub speed: f32,
+    pub radius: f32,
 }
 #[derive(Component)]
 pub struct Frozen {
@@ -81,8 +89,11 @@ fn spawn_enemy(
                 ),
                 YSort { z: 32.0 },
                 Enemy,
-                ChasePlayer { speed: 25.0 },
-                EnemyHealth { health: 1000. },
+                ChasePlayer {
+                    speed: 25.0,
+                    radius: 1000.0,
+                },
+                EnemyHealth { health: 100. },
             ))
             .id();
 
@@ -152,17 +163,19 @@ fn kill_dead_enemies(
         return;
     };
     for (health, transform, entity) in enemy_q.iter() {
-
         if health.health <= 0. {
             commands.entity(entity).insert(AudioPlayer::new(
                 asset_server.load("sounds/snowman_death.ogg"),
             ));
             commands.entity(entity).despawn_recursive();
             commands.spawn((
-                EnemyXp { xp: 10 },
+                EnemyXp { xp: 1.0 },
                 Sprite::from_image(asset_server.load("xp.png")),
                 Transform::from_xyz(transform.translation.x, transform.translation.y, 0.0),
-                ChasePlayer { speed: 100.0 },
+                ChasePlayer {
+                    speed: 100.0,
+                    radius: 80.0,
+                },
             ));
             player_health.hp += 0.2;
         }
@@ -179,6 +192,9 @@ fn chase_player(
     };
     //println!("PlayerPositon coords: {}/{}", player.translation().x, player.translation().y)
     for (mut tf, chase_player) in q.iter_mut() {
+        if tf.translation.distance(player.translation()) > chase_player.radius {
+            continue;
+        }
         let dt = time.delta_secs() * chase_player.speed as f32;
         let dir = (player.translation().truncate() - tf.translation.truncate())
             .normalize()
@@ -187,16 +203,11 @@ fn chase_player(
     }
 }
 
-fn unfreeze(
-    mut commands: Commands,
-    mut q: Query<(&mut Frozen, Entity)>,
-    time: Res<Time>,
-) {
+fn unfreeze(mut commands: Commands, mut q: Query<(&mut Frozen, Entity)>, time: Res<Time>) {
     for (mut frozen, frozen_entity) in q.iter_mut() {
         let dt = time.delta_secs();
         frozen.duration -= dt;
-        if frozen.duration <= 0.0
-        {
+        if frozen.duration <= 0.0 {
             commands.entity(frozen_entity).remove::<Frozen>();
         }
     }
@@ -212,8 +223,7 @@ fn extinguish(
         let dt = time.delta_secs();
         on_fire.duration -= dt;
         enemy_health.health -= stats.fire_dps * dt;
-        if on_fire.duration <= 0.0
-        {
+        if on_fire.duration <= 0.0 {
             commands.entity(on_fire_entity).remove::<OnFire>();
             commands.entity(children[1]).remove::<Blink>();
             commands.entity(children[1]).insert(FlashingTimer {
@@ -232,8 +242,7 @@ fn vunerable_tickdown(
     for (mut vunerable, vunerable_entity) in q.iter_mut() {
         let dt = time.delta_secs();
         vunerable.duration -= dt;
-        if vunerable.duration <= 0.0
-        {
+        if vunerable.duration <= 0.0 {
             commands.entity(vunerable_entity).remove::<Vunerable>();
         }
     }
